+++ src/test/java/ch/liquidmind/inflection/model/external/ViewTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertNotNull;
 
 import org.junit.Test;
 
-import ch.liquidmind.inflection.test.blackbox.BlackboxTestUtility;
+import ch.liquidmind.inflection.test.blackbox.util.BlackboxTestUtility;
 import ch.liquidmind.inflection.test.model.A;
 import ch.liquidmind.inflection.test.model.B1;
