+++ src/test/java/ch/liquidmind/inflection/test/blackbox/util/BlackboxTestUtility.java
@@ -1,4 +1,4 @@
-package ch.liquidmind.inflection.test.blackbox;
+package ch.liquidmind.inflection.test.blackbox.util;
 
 import ch.liquidmind.inflection.loader.TaxonomyLoader;
 import ch.liquidmind.inflection.model.external.Taxonomy;
